Renamed mutual_information module -> information
