+++ depynd/__init__.py
@@ -1,2 +1,2 @@
 __all__ = ['bayesian_networks', 'feature_selection',
-           'markov_networks', 'mutual_information', ]
+           'markov_networks', 'information', ]
