+++ depynd/feature_selection/mifs_.py
@@ -1,5 +1,5 @@
 import numpy as np
-from ..mutual_information import conditional_mutual_information
+from ..information import conditional_mutual_information
 
 
 def mifs(X, y, lamb=0.0, method='knn', options=None):
