+++ depynd/feature_selection/mrmr_.py
@@ -1,5 +1,5 @@
 import numpy as np
-from ..mutual_information import mutual_information
+from ..information import mutual_information
 
 
 def mrmr(X, y, lamb=0.0, method='knn', options=None):
