+++ depynd/information/__init__.py
@@ -1,0 +1,4 @@
+from .information_ import mutual_information, conditional_mutual_information, mimat, cmimat
+
+
+__all__ = ['mutual_information', 'conditional_mutual_information', 'mimat', 'cmimat']
