+++ depynd/information/information_.py
