+++ depynd/information/knn_.py
