+++ depynd/markov_networks/gsmn_.py
@@ -1,5 +1,5 @@
 import numpy as np
-from ..mutual_information import conditional_mutual_information
+from ..information import conditional_mutual_information
 
 
 def gsmn(X, lamb=0.0, method=None, options=None):
