--- depynd/mutual_information/__init__.py
@@ -1,4 +1,0 @@
-from .mutual_information_ import mutual_information, conditional_mutual_information, mimat, cmimat
-
-
-__all__ = ['mutual_information', 'conditional_mutual_information', 'mimat', 'cmimat']
